delete done

+++ Octave_SRM/app/Http/Controllers/CRUD/DController.php
@@ -22,26 +22,34 @@ class DController extends Controller{
 
     public function delete_asset($asset_id)
 {
-
-    $RI = Risk_Identification::where('asset_id', $asset_id);
-    $severity=Severity::where('AoC_id', $RI->AoC_id);
-    $severity->delete();
+    //find RI
+    $RIs = Risk_Identification::where('asset_id', $asset_id)->get();
+    foreach ($RIs as $RI){//for every RI, get severity
+    $severities=Severity::where('AoC_id', $RI->AoC_id)->get();
+        foreach($severities as $severity){//for every severity, delete
+            $severity->delete();
+        }
+    //after that, delete the RI
     $RI->delete();
-
+    }
     //mapping delete
-    $map_h = Map_Human::where('asset_id', $asset_id);
-    $map_h->delete();
-    $map_p = Map_Physical::where('asset_id', $asset_id);
-    $map_p->delete();
-    $map_t = Map_Technical::where('asset_id', $asset_id);
-    $map_t->delete();
+    Map_Human::where('asset_id', $asset_id)->each(function($map_h){
+        $map_h->delete();
+    });
+    Map_Physical::where('asset_id', $asset_id)->each(function($map_p){
+        $map_p->delete();
+    });
+    Map_Technical::where('asset_id', $asset_id)->each(function($map_t){
+        $map_t->delete();
+    });
     //priority delete
-    $priority = Priority::where('asset_id', $asset_id);
+    $priority = Priority::where('asset_id', $asset_id)->first();
+    if ($priority){
     $priority->delete();
-
+    }
     //asset delete
     $asset = Asset::findOrFail($asset_id);
     $asset->delete();
-    return redirect()->route('home')->with('success', 'Asset deleted successfully.');
+    return redirect()->route('home')->with('success', 'Asset and all related records deleted successfully.');
 }
 }
